Initial commit
2/11/2020

Initial commit for TENET v2.0 files. As it stands this data should recapitulate the analysis in "CENPA, MYBL2, and FOXM1 are identified as key transcriptional regulators in lung adenocarcinoma"

Data does not include .rda files necessary for cell type adjustments from TENET v1.0. Additionally we will be looking to further update the method with new information and to streamline the process in later updates.