Update get.diffmeth.regions.R
Bugfix #5:

Another error with the script regarding leukocyte, fibroblast, and sm datasets. Besides loading the dataset, the script had attempted to calculate data from them regardless of settings.

This fix should wrap the addition of the info from those datasets into the dataframe in an if() statement to only run them when the dataset is selected in the settings file.